refactor: 게시글 컴포넌트화
